requirements.txt
playwright==1.40.0
anthropic==0.7.8
supabase==2.0.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
websockets>=12.0
httpx>=0.24.0
wolframalpha==5.0.0
openpyxl==3.1.2
